requirements.txt
altgraph==0.17.2
anyio==3.6.1
asks==3.0.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
docopt==0.6.2
future==0.18.2
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.1
gTTS==2.2.3
h11==0.13.0
httplib2==0.20.4
idna==3.3
Js2Py==0.71
lxml==4.8.0
mega.py==1.0.8
MouseInfo==0.1.3
numpy==1.22.3
opencv-python==4.5.5.64
outcome==1.1.0
packaging==21.3
pathlib==1.0.1
pefile==2021.9.3
Pillow==9.1.0
pipreqs==0.4.11
pipwin==0.5.1
playsound==1.2.2
protobuf==3.20.1
pyarmor==7.4.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/not3v_rnnk8xy/pipwin/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.14.1
PyGetWindow==0.0.9
pyinput==0.3.2
pyinstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.zip
pyinstaller-hooks-contrib==2022.4
pyjsparser==2.7.1
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.6
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==0.1.28
pySmartDL==1.3.4
pyttsx3==2.90
pytweening==1.0.4
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywin32-ctypes==0.2.0
requests==2.26.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
tenacity==5.1.5
trio==0.20.0
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
yarg==0.1.9
